RA-08 Создал адаптер и view holder для item_category.xml. Привязал к recyclerView во фрагменте
