+++ app/src/main/java/com/example/asfoapp/data/recipes.kt
@@ -8,9 +8,6 @@ object STUB {
         Category(3, "Рыба", "Печеная, жареная, сушеная, любая рыба на твой вкус", "fish.png"),
         Category(4, "Супы", "От классики до экзотики: мир в одной тарелке", "soup.png"),
         Category(4, "Супы", "От классики до экзотики: мир в одной тарелке", "soup.png"),
-        Category(4, "Супы", "От классики до экзотики: мир в одной тарелке", "soup.png"),
-        Category(4, "Супы", "От классики до экзотики: мир в одной тарелке", "soup.png"),
-        Category(4, "Супы", "От классики до экзотики: мир в одной тарелке", "soup.png"),
         Category(5, "Салаты", "Хрустящий калейдоскоп под соусом вдохновения", "salad.png")
     )
     fun getCategories(): List<Category> = categories
