+++ app/src/main/java/com/example/asfoapp/ui/categories/CategoriesListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.asfoapp
+package com.example.asfoapp.ui.categories
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.asfoapp.data.STUB
 import com.example.asfoapp.databinding.FragmentCategoriesListBinding
-import com.example.asfoapp.recycler.CategoriesListAdapter
 
 class CategoriesListFragment : Fragment() {
     private var _binding: FragmentCategoriesListBinding? = null
@@ -22,9 +21,7 @@ class CategoriesListFragment : Fragment() {
     ): View? {
         _binding = FragmentCategoriesListBinding.inflate(inflater, container, false)
         val view = binding.root
-
-        val adapter = CategoriesListAdapter(STUB.getCategories())
-        binding.rvCategories.adapter = adapter
+        initRecycler(binding)
 
         return view
     }
@@ -33,4 +30,10 @@ class CategoriesListFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+}
+
+fun initRecycler(binding: FragmentCategoriesListBinding) {
+    val categoriesList = STUB.getCategories()
+    val adapter = CategoriesListAdapter(categoriesList)
+    binding.rvCategories.adapter = adapter
 }
