+++ app/src/main/java/com/example/asfoapp/ui/favorites/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package com.example.asfoapp
+package com.example.asfoapp.ui.favorites
 
 import android.os.Bundle
 import android.view.LayoutInflater
