+++ app/src/main/res/layout/fragment_categories_list.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:background="@color/figma_background_color"
     android:orientation="vertical"
-    tools:context=".CategoriesListFragment">
+    tools:context=".ui.categories.CategoriesListFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
