+++ app/src/main/res/layout/fragment_favorites.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavoritesFragment"
+    tools:context=".ui.favorites.FavoritesFragment"
     android:background="@color/figma_background_color">
 
     <TextView
