+++ app/src/main/res/layout/item_category.xml
@@ -21,7 +21,8 @@
             android:layout_width="match_parent"
             android:layout_height="@dimen/layout_height_categories_card_image"
             android:scaleType="centerCrop"
-            android:src="@drawable/img_error"/>
+            android:src="@drawable/img_error"
+            android:contentDescription="@string/category_image_content_description"/>
 
         <TextView
             android:id="@+id/cardViewTitle"
